more mapper fun, tests and moved utils around
