+++ djangae/contrib/mappers/tests.py
@@ -5,6 +5,7 @@ from django.db import models
 
 from djangae.contrib.mappers.pipes import start_django_orm_mapper
 from djangae.contrib.mappers.queryset import QueryDef
+from djangae.test import process_task_queues
 
 class TestNode(models.Model):
     data = models.CharField(max_length=32)
@@ -31,4 +32,5 @@ class MapReduceTestCase(TestCase):
     def test_all_models_split(self):
         query_def = QueryDef('testapp.TestModel').all()
         start_django_orm_mapper(query_def, 'testapp.urls.test_map_function', 'test_map')
+        process_task_queues()
         self.assertEqual(TestNode.objects.count(), 50)
